-fixed translation data problems -finished book search and book details screens

+++ app/src/main/java/com/ayoprez/sobuu/shared/models/api_models/books_api/Extras.kt
@@ -8,6 +8,6 @@ data class Extras(
     val peopleReadingIt: Int,
     val readingStatus: Int,
     val totalComments: Int,
-    val totalRating: Int,
+    val totalRating: Double,
     val userRating: UserRating?
 )
